Refactor repo interface file structure
